Wraps I/O code in functions
Generally speaking, you don't want side effects like disk access
or network calls to occur when you eval your namespace (since you
tend to reload namespace a lot!). This just wraps up that code
in functions so it is only called when -main is run.

I also just call '(cred)' once in main so we don't incur the cost
of reading the file and parsing it every time we want to use the creds.