Refactor functions to pass in creds file string
It's much easier to test pure functions, so one strategy is to
pull any IO like reading files up as high as possible in the
call stack and then pass the result through pure functions.

You could also test with mocks, but I really try to avoid mocks
as much as possible (and ideally only for higher-level tests that
test interaction between components).

I also fixed a small bug in which we would would match the empty
string for credentials